refactor(infra): consolidate WBTC warp getter, fix deployer addrs, reuse quote signer
Addresses review feedback on the WBTC multi-collateral warp config:
- Merge the staging getter into getWBTCWarpConfig.ts (single file, prod +
  staging) instead of two files.
- Fix the staging owner/deployer addresses, which were placeholder test
  fixtures, to the real Haggis GCP deployer key (mainnet3-haggis-deployer-key):
  0x22EA…C53877e (EVM/TVM) and its ed25519 SVM pubkey DSdGz…pDvR.
- Add haggisDeployerKeyByProtocol (owners.ts), a protocol-keyed map of the
  Haggis deployer addresses for reuse by future staging deploys.
- Extract WARP_QUOTE_SIGNER (warp/consts.ts) and reuse it here plus in the eni
  and eclipse-USDC-stage getters instead of copy-pasting the literal.
- Drop the fabricated second production quote signer; production now uses only
  the canonical quote signer, matching the other FPWR routes.
