+++ typescript/infra/config/environments/mainnet3/warp/consts.ts
@@ -2,3 +2,7 @@
 export const SEALEVEL_WARP_ROUTE_HANDLER_GAS_AMOUNT = 300_000;
 
 export const STARKNET_WARP_ROUTE_HANDLER_GAS_AMOUNT = 5_000_000;
+
+// EIP-712 signer for OffchainQuotedLinearFee quotes on EVM warp-fee routes
+// (Secret Manager: hyperlane-mainnet3-key-quotesigner).
+export const WARP_QUOTE_SIGNER = '0xEd1829805De615eEFC7303766D395Ea0a1B2b04d';
